warnings cleanup
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9080 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/GuidedMissleStrategy.cxx
@@ -32,8 +32,8 @@ void addShotPuff(const float* pos);
 
 static float limitAngle(float a)
 {
-  if (a < -M_PI) a += 2.0f * M_PI;
-  else if (a >= M_PI) a -= 2.0f * M_PI;
+  if (a < -M_PI) a += (float)(2.0 * M_PI);
+  else if (a >= M_PI) a -= (float)(2.0 * M_PI);
   return a;
 }
